modify data type

+++ test.py
@@ -9,8 +9,8 @@ result=os.popen('ping -c %d www.weibo.com'%(int(args[1])))
 doc=result.read()
 pattern=re.compile('time=.*? ')
 tmp = pattern.finditer(doc)
-total=0
+total=0.0
 for item in tmp:
-    time=int(item.group().split('=')[1].rstrip(' '))
+    time=float(item.group().split('=')[1].rstrip(' '))
     total+=time
-print 'averge time is %d'%(total/int(args[1]))
+print 'averge time is %f'%(total/float(args[1]))
